requirements.txt
accelerate
bitsandbytes
llama-index
llama-index-llms-huggingface
llama-index-embeddings-langchain
sentence-transformers
langchain
streamlit
nltk